createCategory,  featchData, deleteCategories, updateCategories
createCategory : makes a post with the category data, this data is entered by the user.

featchData: Get data from API and show user in simple list.

deleteCategories: does a data deletion in the api, when this is done, the flatlist is automatically updated.

updateCategories: Patch the api and update the data and the flatlist, but there is no time when this isn't working and I don't know why.

i created two modals too, one modal for add categories, and other to update categories